+++ NEWS.md
@@ -8,6 +8,8 @@ New language features
   difference between `public` and `export` is that `public` names do not become
   available when `using` a package/module. ([#50105])
 * `ScopedValue` implement dynamic scope with inheritance across tasks ([#50958]).
+* The new macro `Base.Cartesian.@ncallkw` is analogous to `Base.Cartesian.@ncall`,
+  but allows to add keyword arguments to the function call ([#51501]).
 * Support for Unicode 15.1 ([#51799]).
 * A new `AbstractString` type, `AnnotatedString`, is introduced that allows for
   regional annotations to be attached to an underlying string. This type is